fix: remove sqlx dependency from recon and core crates
Modified the store implementations to be interest/model specific as they have different requirements.
Interests don't have values, so those requests no-op (they error if you specify a value other than []).
Removed the sort_key column from the tables as its purpose was interest/model differentiation.
Renamed the tables accordingly and adjusted the queries.
