+++ store/src/sqlite/mod.rs
@@ -1,3 +1,9 @@
+mod interest;
+mod model;
+
+pub use interest::InterestStore;
+pub use model::ModelStore;
+
 use std::{path::Path, str::FromStr};
 
 use sqlx::{
